added an odgi extract test; fix get step position in path; put functions in order; removed redundant import in other test files
