+++ src/subcommand/extract_main.cpp
@@ -4,163 +4,13 @@
 #include <omp.h>
 #include <regex>
 #include "utils.hpp"
-#include "src/algorithms/subgraph/region.hpp"
-#include "src/algorithms/explode.hpp"
+
+#include "src/algorithms/subgraph/extract.hpp"
 
 namespace odgi {
 
     using namespace odgi::subcommand;
 
-    std::pair<step_handle_t, int64_t>
-    get_graph_pos(const odgi::graph_t &graph, const path_handle_t path, const uint64_t &pos) {
-        auto path_end = graph.path_end(path);
-        uint64_t walked = 0;
-        for (step_handle_t s = graph.path_begin(path);
-             s != path_end; s = graph.get_next_step(s)) {
-            handle_t h = graph.get_handle_of_step(s);
-            uint64_t node_length = graph.get_length(h);
-            if (walked + node_length > pos) {
-                return std::make_pair(s, pos - walked);
-            }
-            walked += node_length;
-        }
-#pragma omp critical (cout)
-        std::cerr << "[odgi::get_graph_pos] warning: position " << graph.get_path_name(path) << ":" << pos
-                  << " outside of path" << std::endl;
-        return std::make_pair(path_end, -1);
-    }
-
-    // Create a subpath name
-    string make_subpath_name(const string& path_name, size_t offset, size_t end_offset) {
-        string out_name = path_name + "[" + std::to_string(offset);
-        if (end_offset > 0) {
-            out_name += "-" + std::to_string(end_offset);
-        }
-        out_name += "]";
-        return out_name;
-    }
-
-    /// add subpaths to the subgraph, providing a concatenation of subpaths that are discontiguous over the subgraph
-    /// based on their order in the path position index provided by the source graph
-    /// will clear any path found in both graphs before writing the new steps into it
-    /// if subpath_naming is true, a suffix will be added to each path in the subgraph denoting its offset
-    /// in the source graph (unless the subpath was not cut up at all)
-    void algorithms_add_subpaths_to_subgraph(const graph_t& source, graph_t& subgraph, bool subpath_naming) {
-        auto get_position_of_step =
-                [](const odgi::graph_t& graph, const step_handle_t& step_to_find) {
-                    path_handle_t path = graph.get_path_handle_of_step(step_to_find);
-                    auto path_end = graph.path_end(path);
-                    uint64_t walked = 0;
-                    for (step_handle_t s = graph.path_begin(path); s != path_end; s = graph.get_next_step(s)) {
-                        if (s == step_to_find) {
-                            return walked;
-                        }
-
-                        handle_t h = graph.get_handle_of_step(s);
-                        uint64_t node_length = graph.get_length(h);
-                        walked += node_length;
-                    }
-#pragma omp critical (cout)
-                    std::cerr << "[odgi::get_position_of_step] warning: step " << graph.get_id((graph.get_handle_of_step(step_to_find))) << " in " << graph.get_path_name(path) << " not found" << std::endl;
-
-                    return walked;
-                };
-
-        std::unordered_map<std::string, std::map<uint64_t, handle_t> > subpaths;
-        subgraph.for_each_handle([&](const handle_t& h) {
-            handlegraph::nid_t id = subgraph.get_id(h);
-            if (source.has_node(id)) {
-                handle_t handle = source.get_handle(id);
-                source.for_each_step_on_handle(handle, [&](const step_handle_t& step) {
-                    path_handle_t path = source.get_path_handle_of_step(step);
-                    std::string path_name = source.get_path_name(path);
-                    uint64_t pos = get_position_of_step(source, step);
-                    subpaths[path_name][pos] = source.get_is_reverse(source.get_handle_of_step(step)) ? subgraph.flip(h) : h;
-
-                });
-                /*//return iteratee(step, get_is_reverse(get_handle_of_step(step)) != get_is_reverse(handle), get_position_of_step(step));
-                source.for_each_step_position_on_handle(handle, [&](const step_handle_t& step, const bool& is_rev, const uint64_t& pos) {
-                    path_handle_t path = source.get_path_handle_of_step(step);
-                    std::string path_name = source.get_path_name(path);
-                    subpaths[path_name][pos] = is_rev ? subgraph.flip(h) : h;
-                    return true;
-                });*/
-            }
-        });
-
-        function<path_handle_t(const string&, bool, size_t)> new_subpath =
-                [&subgraph](const string& path_name, bool is_circular, size_t subpath_offset) {
-                    string subpath_name = make_subpath_name(path_name, subpath_offset, 0);
-                    if (subgraph.has_path(subpath_name)) {
-                        subgraph.destroy_path(subgraph.get_path_handle(subpath_name));
-                    }
-                    return subgraph.create_path_handle(subpath_name, is_circular);
-                };
-
-        for (auto& subpath : subpaths) {
-            const std::string& path_name = subpath.first;
-            path_handle_t source_path_handle = source.get_path_handle(path_name);
-            // destroy the path if it exists
-            if (subgraph.has_path(path_name)) {
-                subgraph.destroy_path(subgraph.get_path_handle(path_name));
-            }
-            // create a new path.  give it a subpath name if the flag's on and its smaller than original
-            path_handle_t path;
-            if (!subpath_naming || subpath.second.size() == source.get_step_count(source_path_handle) ||
-                subpath.second.empty()) {
-                path = subgraph.create_path_handle(path_name, source.get_is_circular(source_path_handle));
-            } else {
-                path = new_subpath(path_name, source.get_is_circular(source_path_handle), subpath.second.begin()->first);
-            }
-            for (auto p = subpath.second.begin(); p != subpath.second.end(); ++p) {
-                const handle_t& handle = p->second;
-                if (p != subpath.second.begin() && subpath_naming) {
-                    auto prev = p;
-                    --prev;
-                    const handle_t& prev_handle = prev->second;
-                    // distance from map
-                    size_t delta = p->first - prev->first;
-                    // what the distance should be if they're contiguous depends on relative orienations
-                    size_t cont_delta = subgraph.get_length(prev_handle);
-                    if (delta != cont_delta) {
-                        // we have a discontinuity!  we'll make a new path can continue from there
-                        assert(subgraph.get_step_count(path) > 0);
-                        path = new_subpath(path_name, subgraph.get_is_circular(path), p->first);
-                    }
-                }
-                //fill in the path information
-                subgraph.append_step(path, handle);
-            }
-        }
-    }
-
-    void algorithms_extract_path_range(const graph_t &source, path_handle_t path_handle, int64_t start, int64_t end,
-                                       graph_t &subgraph) {
-
-        // ToDo: could be improved getting and extracting simultaneously (not first get_graph_pos and then extract_path_range)?
-        std::pair<step_handle_t, uint64_t> start_step_and_start_pos = get_graph_pos(source, path_handle, start);
-
-        step_handle_t start_step = start_step_and_start_pos.first;  //source.get_step_at_position(path_handle, start);
-        size_t start_position = start_step_and_start_pos.second;    //source.get_position_of_step(start_step);
-        size_t size_needed = end < 0 ? numeric_limits<size_t>::max() : end - start + 1 + start - start_position;
-        size_t running_length = 0;
-
-        auto path_end = source.path_end(path_handle);
-        for (step_handle_t cur_step = start_step;
-             cur_step != path_end && running_length < size_needed; cur_step = source.get_next_step(cur_step)) {
-
-            handle_t cur_handle = source.get_handle_of_step(cur_step);
-            subgraph.create_handle(source.get_sequence(cur_handle), source.get_id(cur_handle));
-            if (cur_step != start_step) {
-                handle_t prev_handle = source.get_handle_of_step(source.get_previous_step(cur_step));
-                subgraph.create_edge(
-                        subgraph.get_handle(source.get_id(prev_handle), source.get_is_reverse(prev_handle)),
-                        subgraph.get_handle(source.get_id(cur_handle), source.get_is_reverse(cur_handle)));
-            }
-            running_length += source.get_length(cur_handle);
-
-        }
-    }
 
     int main_extract(int argc, char **argv) {
 
@@ -282,7 +132,7 @@ namespace odgi {
                 } else {
                     algorithms::add_connecting_edges_to_subgraph(source, subgraph);
                 }
-                algorithms_add_subpaths_to_subgraph(source, subgraph, false);
+                algorithms::add_subpaths_to_subgraph(source, subgraph, true);
                 //TODO TODO graph.remove_orphan_edges();
                 // Order the mappings by rank. TODO: how do we handle breaks between
                 // different sections of a path with a single name?
@@ -298,7 +148,7 @@ namespace odgi {
                 }
 
                 path_handle_t path_handle = graph.get_path_handle(target.seq);
-                algorithms_extract_path_range(graph, path_handle, target.start, target.end, extract);
+                algorithms::extract_path_range(graph, path_handle, target.start, target.end, extract);
 
                 prep_graph(graph, extract, context_size);
