+++ src/unittest/sort.cpp
@@ -1,7 +1,6 @@
 #include "catch.hpp"
 
 #include <handlegraph/handle_graph.hpp>
-#include <handlegraph/handle_graph.hpp>
 #include <handlegraph/util.hpp>
 #include "odgi.hpp"
 #include "algorithms/topological_sort.hpp"
